Moved test utility function to its own dir
